+++ qtrangeslider/_tests/test_single_value_sliders.py
@@ -1,39 +1,17 @@
 import math
 from contextlib import suppress
+from distutils.version import LooseVersion
 
 import pytest
 
 from qtrangeslider import QDoubleSlider, QLabeledDoubleSlider, QLabeledSlider
 from qtrangeslider._generic_slider import _GenericSlider
+from qtrangeslider.qtcompat import QT_VERSION
 from qtrangeslider.qtcompat.QtCore import QEvent, QPoint, QPointF, Qt
-from qtrangeslider.qtcompat.QtGui import QHoverEvent, QMouseEvent, QWheelEvent
+from qtrangeslider.qtcompat.QtGui import QHoverEvent
 from qtrangeslider.qtcompat.QtWidgets import QSlider, QStyle, QStyleOptionSlider
 
-
-def _mouse_event(pos=QPointF(), type_=QEvent.MouseMove):
-    """Create a mouse event of `type_` at `pos`."""
-    return QMouseEvent(type_, QPointF(pos), Qt.LeftButton, Qt.LeftButton, Qt.NoModifier)
-
-
-def _wheel_event(arc):
-    """Create a wheel event with `arc`."""
-    return QWheelEvent(
-        QPointF(),
-        QPointF(),
-        QPoint(arc, arc),
-        QPoint(arc, arc),
-        Qt.NoButton,
-        Qt.NoModifier,
-        Qt.ScrollBegin,
-        False,
-        Qt.MouseEventSynthesizedByQt,
-    )
-
-
-def _linspace(start, stop, n):
-    h = (stop - start) / (n - 1)
-    for i in range(n):
-        yield start + h * i
+from ._testutil import _linspace, _mouse_event, _wheel_event
 
 
 @pytest.fixture(params=[Qt.Horizontal, Qt.Vertical], ids=["horizontal", "vertical"])
@@ -63,7 +41,8 @@ def sld(request, qtbot, orientation):
     _assert_value_in_range(slider)
     yield slider
     _assert_value_in_range(slider)
-    slider.initStyleOption(QStyleOptionSlider())
+    with suppress(AttributeError):
+        slider.initStyleOption(QStyleOptionSlider())
 
 
 def called_with(*expected_result):
@@ -90,7 +69,7 @@ def test_change_floatslider_range(sld: _GenericSlider, qtbot):
         (BOTH, [called_with(20, 25), called_with(25)], "setMaximum", (25,)),
         ([sld.valueChanged], [called_with(23)], "setValue", (23,)),
     ]:
-        with qtbot.waitSignals(signals, check_params_cbs=checks):
+        with qtbot.waitSignals(signals, check_params_cbs=checks, timeout=500):
             getattr(sld, funcname)(*args)
         _assert_value_in_range(sld)
 
@@ -104,7 +83,7 @@ def test_float_values(sld: _GenericSlider, qtbot):
         (sld.valueChanged, called_with(0.1), "setValue", (0,)),
         (sld.valueChanged, called_with(0.9), "setValue", (1.9,)),
     ]:
-        with qtbot.waitSignal(signals, check_params_cb=checks):
+        with qtbot.waitSignal(signals, check_params_cb=checks, timeout=400):
             getattr(sld, funcname)(*args)
         _assert_value_in_range(sld)
 
@@ -130,26 +109,28 @@ def test_press_move_release(sld: _GenericSlider, qtbot):
     hrect = style.subControlRect(QStyle.CC_Slider, opt, QStyle.SC_SliderHandle)
     handle_pos = _real_sld.mapToGlobal(hrect.center())
 
-    with qtbot.waitSignal(sld.sliderPressed):
+    with qtbot.waitSignal(_real_sld.sliderPressed, timeout=300):
         qtbot.mousePress(_real_sld, Qt.LeftButton, pos=handle_pos)
 
     with suppress(AttributeError):
         assert sld._pressedControl == QStyle.SubControl.SC_SliderHandle
 
-    with qtbot.waitSignals([sld.sliderMoved, _real_sld.valueChanged]):
+    with qtbot.waitSignals(
+        [_real_sld.sliderMoved, _real_sld.valueChanged], timeout=300
+    ):
         shift = (
             QPoint(0, -8) if _real_sld.orientation() == Qt.Vertical else QPoint(8, 0)
         )
         _real_sld.mouseMoveEvent(_mouse_event(handle_pos + shift))
 
-    with qtbot.waitSignal(_real_sld.sliderReleased):
+    with qtbot.waitSignal(_real_sld.sliderReleased, timeout=300):
         qtbot.mouseRelease(_real_sld, Qt.LeftButton, pos=handle_pos)
 
     with suppress(AttributeError):
         assert _real_sld._pressedControl == QStyle.SubControl.SC_None
 
     sld.show()
-    with qtbot.waitSignal(_real_sld.sliderPressed):
+    with qtbot.waitSignal(_real_sld.sliderPressed, timeout=300):
         qtbot.mousePress(_real_sld, Qt.LeftButton, pos=handle_pos)
 
 
@@ -178,8 +159,11 @@ def test_hover(sld: _GenericSlider):
 
 def test_wheel(sld: _GenericSlider, qtbot):
 
+    if type(sld) is QLabeledSlider and LooseVersion(QT_VERSION) < LooseVersion("5.12"):
+        pytest.skip()
+
     _real_sld = getattr(sld, "_slider", sld)
-    with qtbot.waitSignal(sld.valueChanged):
+    with qtbot.waitSignal(sld.valueChanged, timeout=400):
         _real_sld.wheelEvent(_wheel_event(120))
 
     _real_sld.wheelEvent(_wheel_event(0))
@@ -223,7 +207,7 @@ def test_slider_extremes(sld: _GenericSlider, mag, qtbot):
         pytest.skip()
 
     _mag = 10 ** mag
-    with qtbot.waitSignal(sld.rangeChanged):
+    with qtbot.waitSignal(sld.rangeChanged, timeout=400):
         sld.setRange(-_mag, _mag)
     for i in _linspace(-_mag, _mag, 10):
         sld.setValue(i)
