+++ tox.ini
@@ -1,6 +1,6 @@
 # For more information about tox, see https://tox.readthedocs.io/en/latest/
 [tox]
-envlist = py{37,38,39}-{linux,macos,windows}-{pyqt5,pyside2,pyqt6,pyside6}
+envlist = py{37,38,39}-{linux,macos,windows}-{pyqt5,pyside2,pyqt6,pyside6,pyqt511,pyside511}
 toxworkdir=/tmp/.tox
 
 [gh-actions]
@@ -24,6 +24,8 @@ BACKEND =
     pyside2: pyside2
     pyqt6: pyqt6
     pyside6: pyside6
+    pyqt511: pyqt511
+    pyside511: pyside511
 
 [testenv]
 platform =
@@ -33,11 +35,14 @@ platform =
 passenv = CI GITHUB_ACTIONS DISPLAY XAUTHORITY
 deps =
     pytest-xvfb ; sys_platform == 'linux'
+    pyqt511: pyqt5==5.11.*
+    pyside511: pyside2==5.11.*
 extras =
     testing
     pyqt5: pyqt5
     pyside2: pyside2
     pyqt6: pyqt6
     pyside6: pyside6
-commands_pre = pip install -U pytest-qt@git+https://github.com/pytest-dev/pytest-qt.git
-commands = pytest -v --color=yes --cov=qtrangeslider --cov-append --cov-report=xml --cov-report=html {posargs}
+commands_pre =
+    pyqt6,pyside6: pip install -U pytest-qt@git+https://github.com/pytest-dev/pytest-qt.git
+commands = pytest -v --color=yes --cov=qtrangeslider --cov-report=xml {posargs}
